calcWidth must be of int type, not a short
BUG: 175589

+++ khtml/rendering/table_layout.h
@@ -91,7 +91,7 @@ protected:
 	int maxWidth;
 	short effMinWidth;
 	int effMaxWidth;
-	short calcWidth;
+	int calcWidth;
 	bool emptyCellsOnly;
     };
